feat(frontend/map): map view
Created map view which displays the markers from the backend.
The markers are clustered when necessary.
Some preliminary marker- and cluster-icons are included.
The data is currently randomly generated.
The popup-infos are currently implemented as stubs.
The index-route will now display the map.
Changed the body to be at least as high as the screen.