feat(frontend): fix large search result blocking page by lazy loading
The layouting of the tags shown in the shortinfos of every
project/organization occupied the main thread for a significant
amount of time if there were a lot of entries listed.
Instead of immediately showing all entries, we now only load
the first few and add more as the user scrolls along.
This way, only a few tag layouts need to be performed at a time,
making the performance hit unnoticeable.